requirements.txt
filelock==3.15.4
fsspec==2024.6.1
ImageHash==4.3.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.1
pillow==10.4.0
python-dotenv==1.0.1
PyWavelets==1.6.0
scipy==1.13.1
sympy==1.13.2
torch==2.4.0
torchvision==0.19.0
typing_extensions==4.12.2
